requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
elevenlabs==2.6.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
groq==0.26.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
novu==1.14.0
novu-py==1.3.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pillow==11.2.1
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.4
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
yarl==1.20.0
